requirements.txt
Django==4.2.7
django-allauth==0.59.0
django-cors-headers==4.3.0
googletrans==4.0.0-rc1
psycopg2==2.9.9
django-cte==1.3.2
Pillow==10.1.0
discord==2.3.2
gunicorn==19.10.0
django4-background-tasks==1.2.9
pytube==15.0.0
django-recaptcha==4.0.0
beautifulsoup4==4.12.3
channels==4.1.0
daphne==4.1.0
pandas==2.2.1
Send2Trash==1.8.3
openpyxl==3.1.2
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-thread==0.0.1
selenium==4.26.1